feat(utils): deepFlatThenMap -> flattenDeepThenMap 네이밍 변경 및 flattenDeep 추가
