+++ .changeset/lucky-plums-hope.md
@@ -1,0 +1,5 @@
+---
+'@modern-kit/utils': minor
+---
+
+feat(utils): deepFlatThenMap -> flattenDeepThenMap 네이밍 변경 및 flattenDeep 추가 - @ssi02014
