+++ packages/utils/src/array/flatten/index.ts
@@ -1,7 +1,25 @@
-export const flatten = <T, D extends number = 1>(
+/**
+ * @description 중첩 배열을 지정된 깊이까지 평탄화 합니다.
+ *
+ * Array.prototype.flat과 동일하게 동작합니다. 하지만 성능이 더 우수합니다.(약 6~7배)
+ *
+ * @param arr - 평탄화할 중첩 배열입니다.
+ * @param depth - 평탄화할 깊이입니다. 기본값은 1.
+ * @returns 평탄화가 적용된 새로운 배열입니다.
+ *
+ * @example
+ * const arr = [1, [2, 3], [4, [5, 6]]];
+ *
+ * flatten(arr, 1);
+ * // [1, 2, 3, 4, [5, 6]]
+ *
+ * flatten(arr, 2);
+ * // [1, 2, 3, 4, 5, 6]
+ */
+export function flatten<T, D extends number = 1>(
   arr: T[] | readonly T[],
   depth = 1 as D
-) => {
+) {
   const result: FlatArray<T[], D>[] = [];
   const flooredDepth = Math.floor(depth);
 
@@ -17,4 +35,4 @@ export const flatten = <T, D extends number = 1>(
 
   recursive(arr, 0);
   return result;
-};
+}
