+++ packages/utils/src/array/flattenDeep/index.ts
@@ -1,6 +1,6 @@
 import { ExtractNestedArrayType } from '@modern-kit/types';
 import { flatten } from '../../array';
 
-export const flattenDeep = <T>(arr: T[] | readonly T[]) => {
+export function flattenDeep<T>(arr: T[] | readonly T[]) {
   return flatten(arr, Infinity) as ExtractNestedArrayType<T>[];
-};
+}
