+++ packages/utils/src/array/index.ts
@@ -1,10 +1,11 @@
 export * from './chunk';
 export * from './contains';
 export * from './countOccurrencesInArray';
-export * from './deepFlatThenMap';
 export * from './difference';
 export * from './excludeElements';
 export * from './flatten';
+export * from './flattenDeep';
+export * from './flattenDeepThenMap';
 export * from './intersection';
 export * from './intersectionWithDuplicates';
 export * from './partition';
